full update with model comparisons and demo
