+++ zoobot/data_utils/image_datasets.py
@@ -70,9 +70,10 @@ def get_image_dataset(image_paths, file_format, requested_img_size, batch_size, 
     
     assert len(image_paths) > 0
     assert isinstance(image_paths[0], str)
-    logging.info('Images paths to load as dataset: {}'.format(len(image_paths)))
+    logging.info('Image paths to load as dataset: {}'.format(len(image_paths)))
 
     if check_valid_paths:
+        logging.info('Checking if all paths are valid')
         missing_paths = [path for path in image_paths if not os.path.isfile(path)]
         if missing_paths:
             raise FileNotFoundError(f'Missing {len(missing_paths)} images e.g. {missing_paths[0]}')
